requirements.txt
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
Flask==3.0.0
openpyxl==3.1.2
gunicorn==20.1.0
flask_sqlalchemy==2.5.1
flask_migrate==3.1.0
pytest==6.2.4
pytest-flask==1.2.0